ozone-framework-python-server/requirements.txt
asgiref==3.6.0
Babel==2.12.1
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.2
Django==4.1.7
django-cas-ng==4.3.0
django-cors-headers==3.14.0
django-enum-choices==2.1.4
django-extensions==3.2.1
django-filter==23.1
django-method-override==1.0.4
django-nose==1.4.7
django-rest-swagger==2.2.0
django-webpack-loader==0.7.0
djangorestframework==3.14.0
drf-extensions==0.7.1
drf-yasg==1.21.5
httpretty==1.1.4
idna==3.4
imagesize==1.4.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
nose==1.3.7
openapi-codec==1.3.2
packaging==23.0
psycopg2==2.9.5
pycodestyle==2.10.0
Pygments==2.14.0
pyparsing==3.0.9
python-cas==1.6.0
python-dotenv==1.0.0
pytz==2023.2
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.4
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.3
tzdata==2023.2
uritemplate==4.1.1
urllib3==1.26.15
uuid==1.30
waitress==2.1.2
whitenoise==6.4.0
